requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pydantic
python-dotenv
openai>=1.0.0